// static/enums.ts
export enum AGENT_TYPE {
    PARENT = 'PARENT',
    POPUP = 'POPUP',
    WRAPPER = 'WRAPPER'
}

export enum MESSAGE_TYPE {
    INFORM_SOCKET_CONNECTED = 'INFORM_SOCKET_CONNECTED',
    INFORM_SOCKET_DISCONNECTED = 'INFORM_SOCKET_DISCONNECTED',
    ACK_OUTGOING_CALL_START = 'ACK_OUTGOING_CALL_START',
    ACK_OUTGOING_CALL_END = 'ACK_OUTGOING_CALL_END',
    ACK_OUTGOING_CALL_FAIL = 'ACK_OUTGOING_CALL_FAIL',
    ACK_CALL_HOLD = 'ACK_CALL_HOLD',
    ACK_CALL_UNHOLD = 'ACK_CALL_UNHOLD',
    ACK_CALL_MUTE = 'ACK_CALL_MUTE',
    ACK_CALL_UNMUTE = 'ACK_CALL_UNMUTE',
    POPUP_CLOSED = 'POPUP_CLOSED',
    PING_SESSION_DETAILS = 'PING_SESSION_DETAILS',
    PING_POPUP_ALIVE = 'PING_POPUP_ALIVE',
    ACK_SESSION_DETAILS = 'ACK_SESSION_DETAILS',
    REQUEST_OUTGOING_CALL_START = 'REQUEST_OUTGOING_CALL_START',
    REQUEST_OUTGOING_CALL_END = 'REQUEST_OUTGOING_CALL_END',
    REQUEST_CALL_HOLD = 'REQUEST_CALL_HOLD',
    REQUEST_CALL_UNHOLD = 'REQUEST_CALL_UNHOLD',
    REQUEST_CALL_MUTE = 'REQUEST_CALL_MUTE',
    REQUEST_CALL_UNMUTE = 'REQUEST_CALL_UNMUTE',
    REQUEST_SESSION_DETAILS = 'REQUEST_SESSION_DETAILS',
    REQUEST_INCOMING_CALL_START = 'REQUEST_INCOMING_CALL_START',
    REQUEST_INCOMING_CALL_DECLINE = 'REQUEST_INCOMING_CALL_DECLINE',
    REQUEST_INCOMING_CALL_END = 'REQUEST_INCOMING_CALL_END',
    ACK_CALL_MUTE_FAILED = 'ACK_CALL_MUTE_FAILED',
    ACK_CALL_UNMUTE_FAILED = 'ACK_CALL_UNMUTE_FAILED',
    ACK_CALL_UNHOLD_FAILED = 'ACK_CALL_UNHOLD_FAILED', 
    ACK_CALL_HOLD_FAILED = 'ACK_CALL_HOLD_FAILED',
}